EXP-9/generator/requirements.txt
django==4.2.1
django-rest-framework==0.1.0  